ai_backend/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
ultralytics
numpy
requests
python-multipart
